Export fixes
Exporting seems to work properly now. Fixed/added bit flags property. Added spheres for blowup, obstacle, buildit, and force.

+++ Assets/Scripts/FileIO/Loaders/Giz/objects/PanelLoader.cs
@@ -33,7 +33,7 @@ public class PanelLoader : PropertyLoader
         // ### Type ###
         panel.AddProperty(new EnumProperty("Type", bytes[index], Panel.PanelTypes, "", (e) =>
         {
-            panel.UpdatePanelType((int)e.value);
+            panel.UpdatePanelType((byte)e.value);
         }));
         index++;
